Move invoice-channels endpoint from contacts to economy
