+++ core/src/adapters/contacts-adapter/index.ts
@@ -1,10 +1,8 @@
 import { loggedAxios, PaginatedResponse } from '@onecore/utilities'
 import {
   Contact,
-  ChannelLookup,
   GetContactResponseBody,
   GetContactsResponseBody,
-  PostChannelsResponseBody,
 } from '@onecore/contacts/schema'
 
 import { AdapterResult } from '@/adapters/types'
@@ -96,19 +94,6 @@ export const makeContactsAdapter = (contactsServiceUrl: string) => {
         `/contacts/by-email-address/${emailAddress}`
       )
       return listResponse(response)
-    },
-
-    async getInvoiceChannels(
-      contactCodes: string[]
-    ): Promise<AdapterResult<ChannelLookup[], 'unknown'>> {
-      const response = await axios.post<PostChannelsResponseBody>(
-        '/contacts/invoice-channels',
-        {
-          contactCodes,
-        }
-      )
-
-      return { ok: true, data: response.data.content }
     },
   }
 }
