+++ libs/types/src/economy/index.ts
@@ -1,3 +1,4 @@
 export * from './imd'
 export * from './invoice'
 export * from './miscellaneous-invoice'
+export * from './channel-lookup'
