+++ services/contacts/.env.template
@@ -6,7 +6,3 @@ XPAND_DATABASE__PORT= #test
 
 ELASTICSEARCH_LOGGING_HOST=
 APPLICATION_NAME=contacts
-
-STRALFORS__BASE_URL=
-STRALFORS__CLIENT_ID=
-STRALFORS__CLIENT_SECRET=
