+++ services/contacts/src/adapters/contact-adapter.ts
@@ -93,13 +93,4 @@ export interface ContactsRepository {
    * @returns A promise that resolves to an array of Contact objects
    */
   getByEmailAddress: (emailAddress: string) => Promise<Contact[]>
-
-  /**
-   * Retrieves multiple contacts by their contact codes.
-   *
-   * @param contactCodes - The contact codes to search for.
-   *
-   * @returns A promise that resolves to an array of Contact objects
-   */
-  getByContactCodes: (contactCodes: string[]) => Promise<Contact[]>
 }
