--- services/contacts/src/adapters/stralfors/index.ts
@@ -1,1 +1,0 @@
-export { stralforsAdapter } from './stralfors-adapter'
