--- services/contacts/src/adapters/stralfors/stralfors-adapter.ts
@@ -1,159 +1,0 @@
-import axios, { AxiosError, AxiosResponse } from 'axios'
-import {
-  StralforsAccessTokenResponseSchema,
-  StralforsGetChannelLookupResponseSchema,
-  StralforsPostChannelLookupResponseSchema,
-} from './schema'
-
-export interface StralforsAdapterConfig {
-  baseUrl: string
-  clientId: string
-  clientSecret: string
-  retryBackoffMs?: number
-  maxRetries?: number
-}
-
-export interface StralforsAdapter {
-  postChannelLookup: (nationalIdentityNumbers: string[]) => Promise<
-    {
-      error: string | null
-      channel: 'Kivra' | 'Billo' | 'eInvoiceB2C' | 'eInvoiceB2B'
-      matchedCandidates: string[] | null
-    }[]
-  >
-}
-
-export const stralforsAdapter = (
-  config: StralforsAdapterConfig
-): StralforsAdapter => {
-  let accessToken: string | null = null
-
-  async function getAccessToken(): Promise<string> {
-    const response = await axios.post(
-      `${config.baseUrl}/v1/oidc/authorization`,
-      {
-        grant_type: 'client_credentials',
-      },
-      {
-        headers: {
-          'Content-Type': 'application/x-www-form-urlencoded',
-          Authorization: `Basic ${btoa(`${config.clientId}:${config.clientSecret}`)}`,
-        },
-      }
-    )
-
-    return StralforsAccessTokenResponseSchema.parse(response.data).access_token
-  }
-
-  async function makeStralforsRequest(
-    url: string,
-    {
-      method = 'GET',
-      data,
-      headers,
-    }: {
-      method?: string
-      data?: any
-      headers?: any
-    },
-    retryCount: number = 0
-  ): Promise<
-    { data: AxiosResponse; error: null } | { data: null; error: AxiosError }
-  > {
-    if (!accessToken) {
-      accessToken = await getAccessToken()
-    }
-
-    try {
-      const response = await axios({
-        url,
-        method,
-        data,
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-          ...headers,
-        },
-      })
-
-      return { data: response.data, error: null }
-    } catch (err: any) {
-      if (err.status === 401) {
-        if (retryCount > 1) {
-          throw new Error('Could not get valid access token')
-        }
-
-        accessToken = await getAccessToken()
-
-        return makeStralforsRequest(
-          url,
-          { method, data, headers },
-          retryCount + 1
-        )
-      }
-
-      return { data: null, error: err as AxiosError }
-    }
-  }
-
-  async function getChannelLookup(correlationId: string) {
-    const response = await makeStralforsRequest(
-      `${config.baseUrl}/rest/outputmanagement/channellookup/v1/lookup/${correlationId}`,
-      {
-        method: 'GET',
-      }
-    )
-
-    if (response.error?.status === 404) {
-      return null
-    }
-
-    return (
-      StralforsGetChannelLookupResponseSchema.parse(response.data).results ??
-      null
-    )
-  }
-
-  return {
-    async postChannelLookup(nationalIdentityNumbers: string[]) {
-      const response = await makeStralforsRequest(
-        `${config.baseUrl}/rest/outputmanagement/channellookup/v1/lookup`,
-        {
-          method: 'POST',
-          data: {
-            channels: ['Kivra', 'Billo', 'eInvoiceB2C'],
-            candidates: nationalIdentityNumbers,
-          },
-        }
-      )
-
-      if (response.error) {
-        throw new Error()
-      }
-
-      const correlationId = StralforsPostChannelLookupResponseSchema.parse(
-        response.data
-      ).correlationId
-
-      const pollGetChannelLookup = async (retryCount: number = 0) => {
-        const result = await getChannelLookup(correlationId)
-
-        if (!result) {
-          if (retryCount > (config.maxRetries ?? 10)) {
-            throw new Error(
-              'Failed to get a response from Strålfors API after 10 retries'
-            )
-          }
-
-          await new Promise((resolve) =>
-            setTimeout(resolve, config.retryBackoffMs ?? 500)
-          )
-          return pollGetChannelLookup(retryCount + 1)
-        }
-
-        return result
-      }
-
-      return pollGetChannelLookup()
-    },
-  }
-}
