+++ services/contacts/src/adapters/xpand/query.ts
@@ -208,7 +208,6 @@ export const contactsQuery = () => {
   let wildcards: string[] = []
   let contactType: ContactTypeFilter = 'any'
   let objectKeyFilter: ContactCode[] = []
-  let contactCodesFilter: ContactCode[] = []
 
   /**
    * Applies the base cmctc-only criteria/filters to a knex
@@ -237,10 +236,6 @@ export const contactsQuery = () => {
 
     if (objectKeyFilter.length) {
       qb.andWhereRaw(...trimInCriteria('cmctc.keycmobj', objectKeyFilter))
-    }
-
-    if (contactCodesFilter.length) {
-      qb.andWhereRaw(...trimInCriteria('cmctc.cmctckod', contactCodesFilter))
     }
 
     Object.keys(wheres).forEach((key) => {
@@ -320,14 +315,6 @@ export const contactsQuery = () => {
      */
     withObjectKeyIn(objectKeys: ObjectKey[]) {
       objectKeyFilter = objectKeys
-      return this
-    },
-
-    /**
-     * Adds a WHERE-criteria for multiple contact codes.
-     */
-    withContactCodeIn(contactCodes: ContactCode[]) {
-      contactCodesFilter = contactCodes
       return this
     },
