--- services/contacts/src/domain/channel-lookup.ts
@@ -1,4 +1,0 @@
-import { ChannelLookupSchema } from '@src/services/contacts-service/schema'
-import z from 'zod'
-
-export type ChannelLookup = z.infer<typeof ChannelLookupSchema>
