+++ services/contacts/src/domain/index.ts
@@ -1,2 +1,1 @@
 export type * from './contact'
-export type * from './channel-lookup'
