+++ services/contacts/src/services/contacts-service/api-types.ts
@@ -2,7 +2,6 @@ import z from 'zod'
 import {
   GetContactResponseBodySchema,
   GetContactsResponseBodySchema,
-  PostChannelsResponseBodySchema,
 } from './schema'
 
 export type GetContactResponseBody = z.infer<
@@ -11,8 +10,4 @@ export type GetContactResponseBody = z.infer<
 
 export type GetContactsResponseBody = z.infer<
   typeof GetContactsResponseBodySchema
->
-
-export type PostChannelsResponseBody = z.infer<
-  typeof PostChannelsResponseBodySchema
 >
