+++ services/contacts/src/services/contacts-service/schema.ts
@@ -97,18 +97,3 @@ export const GetContactsResponseBodySchema =
       contacts: z.array(ContactSchema),
     }),
   })
-
-export const PostChannelsRequestBodySchema = z.object({
-  contactCodes: z.string().array(),
-})
-
-export const ChannelLookupSchema = z.object({
-  channel: z.enum(['Kivra', 'Billo', 'eInvoiceB2C', 'eInvoiceB2B']),
-  matchedCandidates: z.string().array().nullable(),
-  error: z.string().nullable(),
-})
-
-export const PostChannelsResponseBodySchema =
-  ONECoreHateOASResponseBodySchema.extend({
-    content: ChannelLookupSchema.array(),
-  })
