+++ services/economy/.env.template
@@ -32,5 +32,9 @@ TENFAST__BASE_URL=
 TENFAST__API_KEY=
 TENFAST__COMPANY_ID=
 
+STRALFORS__BASE_URL=
+STRALFORS__CLIENT_ID=
+STRALFORS__CLIENT_SECRET=
+
 ELASTICSEARCH_LOGGING_HOST=http://localhost:9208
 APPLICATION_NAME=economy
