+++ services/economy/src/services/invoice-service/adapters/stralfors/schema.ts
