+++ services/economy/src/services/invoice-service/service.ts
@@ -22,6 +22,7 @@ import {
   getRentalInvoices,
   getInvoiceRows as getXpandInvoiceRows,
   getBatchTotalAmount as getXpandBatchTotalAmount,
+  getContacts,
 } from './adapters/xpand-db-adapter'
 import {
   CounterPartCustomers,
@@ -57,6 +58,7 @@ import {
   getInvoicesForTenant,
   getTenantByContactCode,
 } from '@src/common/adapters/tenfast/tenfast-adapter'
+import { postChannelLookup } from './adapters/stralfors/stralfors-adapter'
 
 const createRoundOffRow = async (
   invoice: InvoiceData,
@@ -1048,4 +1050,22 @@ const enrichInvoiceRowsWithText = async (
       invoiceRowText: article?.label ?? null,
     }
   })
+}
+
+export const stralforsPostChannelLookup = async (contactCodes: string[]) => {
+  const contacts = await getContacts(contactCodes)
+
+  // TODO We currently do not have any test contacts in the Strålfors test environment, so we cannot test with actual contacts
+  // const nationalIdentityNumbers = contacts.map((c) => c.nationalRegistrationNumber)
+  const nationalIdentityNumbers = [
+    '191212121212',
+    '198112172385',
+    '197102125866',
+    '197701032380',
+    '192112039223',
+    '198110292383',
+    '198903092388',
+  ]
+
+  return postChannelLookup(nationalIdentityNumbers)
 }
